FINALLY handle last induction date update logic
In my effort to focus solely on updating the schools that have recently submitted their chapter reports, I used the following logic. While this approach has worked successfully for Zone 1, a potential issue may arise where school names in both files are similar but not an exact match, a problem I encountered early on and resolved using fuzzy matching. Since this hasn’t posed a problem yet and I need to update the rest of the columns with limited time remaining, I will keep this line as it is. If it causes issues later or once I have completed everything else, I will address this potential bug at the end.